requirements.txt
addict==2.4.0
easydict==1.10
einops==0.6.1
einshape==1.0
ffmpeg-python==0.2.0
gdown==4.7.1
gin-config==0.4.0
numpy==1.22.4
open3d==0.15.2
opencv-python==4.8.0.76
Pillow==9.4.0
pycocotools==2.0.7
rich==13.5.3
scikit-image==0.19.2
scikit-learn==1.3.0
scipy==1.10.1
timm==0.9.7
tqdm==4.66.1
transformers==4.33.1
trimesh==3.23.5
plotly
flax==0.4.2
chex==0.1.6
optax==0.1.7
lpips==0.1.4

tensorflow-cpu==2.8.0
tensorflow-graphics==2021.12.3
tensorflow-hub==0.12.0
tensorboard==2.8.0